format debugger errors in the vscode extension layer
