+++ source/vscode/src/circuit.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { escapeHtml } from "markdown-it/lib/common/utils.mjs";
 import {
   type CircuitData,
   ICompilerWorker,
@@ -24,8 +23,8 @@ import {
 } from "./telemetry";
 import { getRandomGuid } from "./utils";
 import { sendMessageToPanel } from "./webviewPanel";
-import { ICircuitConfig, IPosition } from "../../npm/qsharp/lib/web/qsc_wasm";
-import { basename, loadCompilerWorker } from "./common";
+import { ICircuitConfig } from "../../npm/qsharp/lib/web/qsc_wasm";
+import { errorsToHtml, loadCompilerWorker } from "./common";
 
 const compilerRunTimeoutMs = 1000 * 60 * 5; // 5 minutes
 
@@ -405,46 +404,6 @@ function hasAdaptiveComplianceOrUnsupportedRirError(errors: IQSharpError[]) {
   return hasResultComparisonError;
 }
 
-/**
- * Formats an array of compiler/runtime errors into HTML to be presented to the user.
- *
- * @param errors The list of errors to format.
- * @returns The HTML formatted errors, to be set as the inner contents of a container element.
- */
-function errorsToHtml(errors: IQSharpError[]) {
-  let errorHtml = "";
-  for (const error of errors) {
-    const { document, diagnostic: diag, stack: rawStack } = error;
-
-    const location = documentHtml(false, document, diag.range.start);
-    const message = escapeHtml(`(${diag.code}) ${diag.message}`).replace(
-      /\n/g,
-      "<br/><br/>",
-    );
-
-    errorHtml += `<p>${location}: ${message}<br/></p>`;
-
-    if (rawStack) {
-      const stack = rawStack
-        .split("\n")
-        .map((l) => {
-          // Link-ify the document names in the stack trace
-          const match = l.match(/^(\s*)at (.*) in (.*):(\d+):(\d+)/);
-          if (match) {
-            const [, leadingWs, callable, doc] = match;
-            return `${leadingWs}at ${escapeHtml(callable)} in ${documentHtml(false, doc)}`;
-          } else {
-            return l;
-          }
-        })
-
-        .join("\n");
-      errorHtml += `<br/><pre>${stack}</pre>`;
-    }
-  }
-  return errorHtml;
-}
-
 export function updateCircuitPanel(
   targetProfile: string,
   projectName: string,
@@ -479,57 +438,4 @@ export function updateCircuitPanel(
     props,
   };
   sendMessageToPanel({ panelType: "circuit", id: panelId }, reveal, message);
-}
-
-/**
- * If the input is a URI, turns it into a document open link.
- * Otherwise returns the HTML-escaped input
- */
-function documentHtml(
-  customCommand: boolean,
-  maybeUri: string,
-  position?: IPosition,
-) {
-  try {
-    // If the error location is a document URI, create a link to that document.
-    // We use the `vscode.open` command (https://code.visualstudio.com/api/references/commands#commands)
-    // to open the document in the editor.
-    // The line and column information is displayed, but are not part of the link.
-    //
-    // At the time of writing this is the only way we know to create a direct
-    // link to a Q# document from a Web View.
-    //
-    // If we wanted to handle line/column information from the link, an alternate
-    // implementation might be having our own command that navigates to the correct
-    // location. Then this would be a link to that command instead. Yet another
-    // alternative is to have the webview pass a message back to the extension.
-    const uri = Uri.parse(maybeUri, true);
-    const fsPath = escapeHtml(basename(uri.path) ?? uri.fsPath);
-    const lineColumn = position
-      ? escapeHtml(`:${position.line + 1}:${position.character + 1}`)
-      : "";
-
-    const locations = [
-      {
-        source: uri,
-        span: {
-          start: position,
-          end: position,
-        },
-      },
-    ];
-
-    const args = customCommand && position ? [locations] : [uri];
-    const openCommand =
-      customCommand && position ? "qsharp-vscode.gotoLocations" : "vscode.open";
-
-    const argsStr = encodeURIComponent(JSON.stringify(args));
-    const openCommandUri = Uri.parse(`command:${openCommand}?${argsStr}`, true);
-    const title = `${fsPath}${lineColumn}`;
-    return `<a href="${openCommandUri}">${title}</a>`;
-  } catch {
-    // Likely could not parse document URI - it must be a project level error
-    // or an error from stdlib, use the document name directly
-    return escapeHtml(maybeUri);
-  }
 }
